Video Upload Updates
Added react-player for displaying video files.
Updated to the latest version of react with create-react-app.
Updated version of leaflet and added the new script and style tags for it in index.html.
Fixed an issue with error toasts not showing properly.
Added a new tab for video uploading in the location editor.
Added an API endpoint for uploading video files to Cloudinary.